quest_1/rag_app/requirements_ui.txt
streamlit==1.35.0
streamlit-aggrid==0.3.4.post3
streamlit-antd-components==0.3.1
streamlit-chatbox==1.1.12.post2
streamlit-feedback==0.1.3
streamlit-modal==0.1.0
streamlit-option-menu==0.3.12
python-dotenv==1.0.1
fastapi==0.109.2
langchain==0.2.1
langchain-community==0.2.4
langchain-core==0.2.5
langchain-experimental==0.0.60
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langchainhub==0.1.18
langdetect==1.0.9
langsmith==0.1.77
chardet==5.2.0
retrying==1.3.3
sse-starlette==1.8.2
streamlit_cookies_manager